+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### [13.14.0](https://kaos.sh/ek/13.14.0)
 
+- `[fmtutil/filetree]` Added experimental package for printing file tree
 - `[fmtc]` Added methods `Printfn`, `Fprintfn`, `Sprintfn`, and `LPrintfn`
 - `[fmtutil/table]` Improved support of data with ANSI escape sequences
 - `[fmtc]` `NameColor` marked as deprecated (_use method `AddColor` instead_)