+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `[fmtutil/filetree]` Added experimental package for printing file tree
 - `[fmtc]` Added methods `Printfn`, `Fprintfn`, `Sprintfn`, and `LPrintfn`
+- `[req]` Added custom encoder for arrays in `Query`
 - `[fmtutil/table]` Improved support of data with ANSI escape sequences
 - `[fmtc]` `NameColor` marked as deprecated (_use method `AddColor` instead_)